Move checkAudioAvailability to audio.ts
